Добавляем опцию, которая только на СП есть, чтобы сурсмапы не генерились для старых дистрибутивов, где есть debug https://online.sbis.ru/opendoc.html?guid=b133b623-ba00-44da-bb2c-2ee8ae894454

+++ Tasks/uglify.js
@@ -10,15 +10,18 @@ var path = require('path');
 
 module.exports = function uglifyJsTask(grunt) {
     grunt.registerMultiTask('uglify', 'Задача минификации JS', function() {
-        var getAvailableFiles = function (filesArray) {
-            return filesArray.filter(function (filepath) {
-                if (!grunt.file.exists(filepath)) {
-                    grunt.log.warn('Source file ' + chalk.cyan(filepath) + ' not found');
-                    return false;
-                }
-                return true;
-            });
-        };
+        var
+            getAvailableFiles = function (filesArray) {
+                return filesArray.filter(function (filepath) {
+                    if (!grunt.file.exists(filepath)) {
+                        grunt.log.warn('Source file ' + chalk.cyan(filepath) + ' not found');
+                        return false;
+                    }
+                    return true;
+                });
+            },
+            inclReplace =  (grunt.option('includes') !== undefined) ? grunt.option('includes') : true;
+
         // Iterate over all src-dest file pairs.
         this.files.forEach(function (currentFile) {
             var availableFiles = getAvailableFiles(currentFile.src);
@@ -32,20 +35,23 @@ module.exports = function uglifyJsTask(grunt) {
                     data = fs.readFileSync(currentPath, 'utf8'),
                     sourceJSPath = currentPath.replace(/\.js$/, '.source.js'),
                     sourceMapPath = `${currentPath}.map`,
-                    dataObject = {};
-
-                dataObject[path.basename(sourceJSPath)] = data;
-                fs.writeFileSync(sourceJSPath, data);
-
-                try {
-                    let minified = UglifyJS.minify(dataObject, {
+                    dataObject = {},
+                    minifyOptions = {
                         mangle: {
                             eval: true
-                        },
-                        sourceMap: {
-                            url: path.basename(sourceMapPath)
                         }
-                    });
+                    };
+                if (inclReplace) {
+                    fs.writeFileSync(sourceJSPath, data);
+                    minifyOptions.sourceMap = {
+                        url: path.basename(sourceMapPath)
+                    };
+
+                }
+                dataObject[path.basename(sourceJSPath)] = data;
+
+                try {
+                    let minified = UglifyJS.minify(dataObject, minifyOptions);
                     if (!minified.error && minified.code) {
                         data = minified;
                     }
@@ -53,7 +59,9 @@ module.exports = function uglifyJsTask(grunt) {
                     grunt.log.warn(`Error while minifiing js! ${minerr.message}, in file: ${fullPath}`);
                 }
                 fs.writeFileSync(currentPath, data.code);
-                fs.writeFileSync(sourceMapPath, data.map);
+                if (inclReplace) {
+                    fs.writeFileSync(sourceMapPath, data.map);
+                }
                 grunt.log.ok(`File ${currentPath} successfully minified`);
             });
         });
